requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
atpublic==4.1.0
attrs==23.2.0
backoff==2.2.1
bidict==0.23.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
comm==0.2.2
datasets==2.14.7
debugpy==1.8.1
decorator==5.1.1
dill==0.3.7
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
dspy==0.1.5
dspy-ai==2.4.5
email_validator==2.2.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
ibis-framework==9.1.0
idna==3.7
instructor==1.3.2
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.1
joblib==1.3.2
jupyter_client==8.6.2
jupyter_core==5.7.2
langfuse==2.35.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.33.0
optuna==3.6.1
orjson==3.10.5
packaging==23.2
pandas==2.2.2
parso==0.8.4
parsy==2.1
pexpect==4.9.0
platformdirs==4.2.2
polars==0.20.31
prompt_toolkit==3.0.47
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlglot==25.1.0
stack-data==0.6.3
starlette==0.37.2
tenacity==8.3.0
tokenizers==0.19.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
